create a LawType component
